# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Partner Watch API
  version: 0.1.0
  description: API contract under development. Undocumented routes are not public API.
servers:
  - url: https://watch.example.com
    description: Example only; the actual origin is supplied through PW_PUBLIC_URL.
paths:
...
  /v1/enrollments:
    post:
      operationId: enrollDevice
      description: One-time enrollment. Invalid requests deliberately return 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentRequest"
      responses:
        "201":
          description: Device enrolled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentResponse"
        "400":
          description: Request JSON or field format is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Enrollment endpoint or invitation was not found.
  /v1/images:
    post:
      operationId: uploadImage
      description: Uploads a JPEG for the other device in the pair. Images expire after one hour.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
              maxLength: 10485760
      responses:
        "201":
          description: Image stored temporarily.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedImage"
        "400":
          description: The body is not a valid JPEG or exceeds the pixel limit.
        "401":
          description: Device credential is invalid.
        "413":
          description: Image exceeds 10 MiB.
        "415":
          description: Content-Type is not image/jpeg.
  /v1/images/{imageId}:
    get:
      operationId: downloadImage
      description: Returns an image once to the paired device and immediately removes it from the server.
      security:
        - bearerAuth: []
      parameters:
        - name: imageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: JPEG image. A successful claim consumes the image.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "401":
          description: Device credential is invalid.
        "404":
          description: Image does not exist, expired, was already retrieved, or belongs to another pair/device.
  /v1/events:
    get:
      operationId: connectEvents
      description: Upgrades to a WebSocket that delivers events for the authenticated device.
      security:
        - bearerAuth: []
      responses:
        "101":
          description: WebSocket connection established.
        "401":
          description: Device credential is invalid.
  /v1/capture-requests:
    post:
      operationId: createCaptureRequest
      description: Creates a request for the other device in the pair. Requests expire after one minute.
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Capture request accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureRequest"
        "401":
          description: Device credential is invalid.
        "409":
          description: The paired device is not enrolled.
        "429":
          description: Limit of one request per 10 seconds or 60 per hour was exceeded.
  /v1/capture-requests/{requestId}/result:
    post:
      operationId: reportCaptureResult
      description: Reports a successful upload or a classified failure from the requested device.
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureResult"
      responses:
        "204":
          description: Result accepted and delivered to the requester when connected.
        "400":
          description: Result fields are invalid.
        "401":
          description: Device credential is invalid.
        "404":
          description: Request does not exist, expired, completed, or belongs to another device.
  /v1/capture-requests/{requestId}:
    get:
      operationId: getCaptureRequest
      description: Returns a request to its requester, including its terminal result for power-saving clients.
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current request state.
        "404":
          description: Request does not exist or belongs to another device.
  /v1/status-requests:
    post:
      operationId: createStatusRequest
      description: Requests the paired device's currently enabled status fields. Limited to once per minute and 20 times per hour.
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Status request accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusRequest"
        "409":
          description: The paired device is not enrolled.
        "429":
          description: Status request rate limit was exceeded.
  /v1/status-requests/pending:
    get:
      operationId: listPendingStatusRequests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unexpired requests targeting this device.
  /v1/status-requests/{requestId}/result:
    post:
      operationId: reportStatusResult
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusResult"
      responses:
        "204":
          description: Partial status result accepted.
        "404":
          description: Request does not exist, expired, completed, or belongs to another device.
  /v1/partner-status:
    get:
      operationId: getPartnerStatus
      description: Returns the paired device's latest unexpired snapshot.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Latest snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusSnapshot"
        "404":
          description: No unexpired snapshot exists.
  /v1/device-status/{field}:
    delete:
      operationId: clearOwnStatusField
      description: Clears one field from the authenticated device's currently stored snapshot.
      security:
        - bearerAuth: []
      parameters:
        - name: field
          in: path
          required: true
          schema:
            type: string
            enum: [battery, location]
      responses:
        "204":
          description: Snapshot deleted or did not exist.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    EnrollmentRequest:
      type: object
      additionalProperties: false
      required: [invitationToken, deviceName, publicKey]
      properties:
        invitationToken:
          type: string
          minLength: 43
          maxLength: 128
        deviceName:
          type: string
          minLength: 1
          maxLength: 80
        publicKey:
          type: string
          description: Unpadded base64url-encoded X.509 SubjectPublicKeyInfo for a P-256 ECDSA key.
    EnrollmentResponse:
      type: object
      additionalProperties: false
      required: [deviceId, pairId, slot, credential]
      properties:
        deviceId:
          type: string
        pairId:
          type: string
        slot:
          type: integer
          minimum: 1
          maximum: 2
        credential:
          type: string
          description: Device credential returned once at enrollment.
    UploadedImage:
      type: object
      additionalProperties: false
      required: [imageId, createdAt, expiresAt]
      properties:
        imageId:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    CaptureRequest:
      type: object
      additionalProperties: false
      required: [requestId, status, createdAt, expiresAt]
      properties:
        requestId:
          type: string
        status:
          type: string
          const: PENDING
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    CaptureResult:
      type: object
      additionalProperties: false
      required: [status, imageId, failure]
      properties:
        status:
          type: string
          enum: [READY, FAILED]
        imageId:
          type: string
          description: Required for READY and empty for FAILED.
        failure:
          type: string
          enum: ["", DISABLED, SERVICE_UNAVAILABLE, LOCKED, CAPTURE_PROTECTED, INTERNAL_ERROR]
          description: Empty for READY and classified for FAILED.
    StatusRequest:
      type: object
      additionalProperties: false
      required: [requestId, status, createdAt, expiresAt]
      properties:
        requestId:
          type: string
        status:
          type: string
          const: PENDING
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    BatteryStatus:
      type: object
      additionalProperties: false
      required: [status, percent, chargingState]
      properties:
        status:
          type: string
          enum: [AVAILABLE, DISABLED]
        percent:
          type: integer
          minimum: 0
          maximum: 100
        chargingState:
          type: string
          enum: ["", CHARGING, DISCHARGING, FULL, NOT_CHARGING, UNKNOWN]
    StatusResult:
      type: object
      additionalProperties: false
      required: [battery, location]
      properties:
        battery:
          $ref: "#/components/schemas/BatteryStatus"
        location:
          $ref: "#/components/schemas/LocationStatus"
    LocationStatus:
      type: object
      additionalProperties: false
      required: [status, latitude, longitude, accuracyMeters, observedAt, source]
      properties:
        status:
          type: string
          enum: [AVAILABLE, DISABLED, PERMISSION_DENIED, SETTING_OFF, UNAVAILABLE, TIMEOUT]
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        accuracyMeters:
          type: number
          minimum: 0
        observedAt:
          type: string
          format: date-time
        source:
          type: string
          enum: ["", FRESH, LAST_KNOWN]
    StatusSnapshot:
      type: object
      additionalProperties: false
      required: [deviceId, battery, location, reportedAt, expiresAt]
      properties:
        deviceId:
          type: string
        battery:
          $ref: "#/components/schemas/BatteryStatus"
        location:
          $ref: "#/components/schemas/LocationStatus"
        reportedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
...
    Error:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code:
          type: string