# server/internal/httpapi/handler.go
package httpapi

import (
	"bytes"
	"context"
	"crypto/ecdsa"
	"crypto/elliptic"
	"crypto/x509"
	"encoding/base64"
	"encoding/json"
	"errors"
	"fmt"
	"image"
	_ "image/jpeg"
	"io"
	"log/slog"
	"net/http"
	"strings"
	"time"
	"unicode/utf8"

	"github.com/coder/websocket"
	"github.com/coder/websocket/wsjson"
	"github.com/kowerkoint/partner-watch/server/internal/store"
)

type healthResponse struct {
	Status string `json:"status"`
}

type Enroller interface {
	EnrollDevice(ctx context.Context, invitationToken, deviceName, publicKey string) (store.Enrollment, error)
}

type Authenticator interface {
	AuthenticateDevice(ctx context.Context, credential string) (string, error)
}

type FCMTokenStore interface {
	SetFCMToken(context.Context, string, string) error
}
type WakeupSender interface {
	SendWakeup(context.Context, string, string) error
}

type ImageStore interface {
	SaveImage(ctx context.Context, deviceID string, data []byte, width, height int) (store.Image, error)
	TakeImage(ctx context.Context, deviceID, imageID string) (store.Image, error)
}

type CaptureStore interface {
	CreateCaptureRequest(ctx context.Context, requesterDeviceID string) (store.CaptureRequest, error)
	CompleteCaptureRequest(ctx context.Context, targetDeviceID, requestID, status, imageID, failure string) (store.CaptureRequest, error)
}
type CaptureReader interface {
	CaptureRequestForRequester(context.Context, string, string) (store.CaptureRequest, error)
}
type PendingCaptureStore interface {
	PendingCaptureRequests(ctx context.Context, deviceID string) ([]store.CaptureRequest, error)
}
type StatusStore interface {
	CreateStatusRequest(context.Context, string) (store.StatusRequest, error)
	PendingStatusRequests(context.Context, string) ([]store.StatusRequest, error)
	CompleteStatusRequest(context.Context, string, string, store.BatteryReport, store.LocationReport) (store.StatusRequest, error)
	PartnerStatusSnapshot(context.Context, string) (store.StatusSnapshot, error)
	ClearStatusField(context.Context, string, string) error
}

type enrollmentRequest struct {
	InvitationToken string `json:"invitationToken"`
	DeviceName      string `json:"deviceName"`
	PublicKey       string `json:"publicKey"`
}

type enrollmentResponse struct {
	DeviceID   string `json:"deviceId"`
	PairID     string `json:"pairId"`
	Slot       int    `json:"slot"`
	Credential string `json:"credential"`
}

type errorResponse struct {
	Code string `json:"code"`
}

type imageResponse struct {
	ImageID   string    `json:"imageId"`
	CreatedAt time.Time `json:"createdAt"`
	ExpiresAt time.Time `json:"expiresAt"`
}

type captureRequestResponse struct {
	RequestID string    `json:"requestId"`
	Status    string    `json:"status"`
	CreatedAt time.Time `json:"createdAt"`
	ExpiresAt time.Time `json:"expiresAt"`
}
type pendingCaptureResponse struct {
	Requests []captureRequestResponse `json:"requests"`
}

type fcmTokenRequest struct {
	Token string `json:"token"`
}
type captureResultRequest struct {
	Status  string `json:"status"`
	ImageID string `json:"imageId"`
	Failure string `json:"failure"`
}
type batteryResultRequest struct {
	Status        string `json:"status"`
	Percent       int    `json:"percent"`
	ChargingState string `json:"chargingState"`
}
type statusResultRequest struct {
	Battery  batteryResultRequest  `json:"battery"`
	Location locationResultRequest `json:"location"`
}
type locationResultRequest struct {
	Status         string    `json:"status"`
	Latitude       float64   `json:"latitude"`
	Longitude      float64   `json:"longitude"`
	AccuracyMeters float64   `json:"accuracyMeters"`
	ObservedAt     time.Time `json:"observedAt"`
	Source         string    `json:"source"`
}
type statusSnapshotResponse struct {
	DeviceID   string                `json:"deviceId"`
	Battery    batteryResultRequest  `json:"battery"`
	Location   locationResultRequest `json:"location"`
	ReportedAt time.Time             `json:"reportedAt"`
	ExpiresAt  time.Time             `json:"expiresAt"`
}

const maxImageBytes = 10 * 1024 * 1024
const maxImagePixels = 5_000_000

func NewHandler(enroller Enroller, sender ...WakeupSender) http.Handler {
	return newHandler(enroller, newEventHub(), sender...)
}

func newHandler(enroller Enroller, events *eventHub, sender ...WakeupSender) http.Handler {
	mux := http.NewServeMux()
	mux.HandleFunc("GET /healthz", handleHealth)
	mux.HandleFunc("POST /v1/enrollments", handleEnrollment(enroller))
	authenticator, authenticated := enroller.(Authenticator)
	if images, ok := enroller.(ImageStore); ok && authenticated {
		mux.HandleFunc("POST /v1/images", authenticate(authenticator, handleImageUpload(images)))
		mux.HandleFunc("GET /v1/images/{imageID}", authenticate(authenticator, handleImageDownload(images)))
	}
	if captures, ok := enroller.(CaptureStore); ok && authenticated {
		mux.HandleFunc("GET /v1/events", handleEvents(authenticator, events))
		mux.HandleFunc("POST /v1/capture-requests", authenticate(authenticator, handleCaptureRequest(captures, events, sender...)))
		mux.HandleFunc("POST /v1/capture-requests/{requestID}/result", authenticate(authenticator, handleCaptureResult(captures, events)))
	}
	if pending, ok := enroller.(PendingCaptureStore); ok && authenticated {
		mux.HandleFunc("GET /v1/capture-requests/pending", authenticate(authenticator, handlePendingCaptureRequests(pending)))
	}
	if reader, ok := enroller.(CaptureReader); ok && authenticated {
		mux.HandleFunc("GET /v1/capture-requests/{requestID}", authenticate(authenticator, handleCaptureRequestStatus(reader)))
	}
	if tokens, ok := enroller.(FCMTokenStore); ok && authenticated {
		mux.HandleFunc("POST /v1/device/fcm-token", authenticate(authenticator, handleFCMToken(tokens)))
	}
	if statuses, ok := enroller.(StatusStore); ok && authenticated {
		mux.HandleFunc("POST /v1/status-requests", authenticate(authenticator, handleStatusRequest(statuses, events, sender...)))
		mux.HandleFunc("GET /v1/status-requests/pending", authenticate(authenticator, handlePendingStatusRequests(statuses)))
		mux.HandleFunc("POST /v1/status-requests/{requestID}/result", authenticate(authenticator, handleStatusResult(statuses, events)))
		mux.HandleFunc("GET /v1/partner-status", authenticate(authenticator, handlePartnerStatus(statuses)))
		mux.HandleFunc("DELETE /v1/device-status/{field}", authenticate(authenticator, handleClearStatus(statuses)))
		mux.HandleFunc("DELETE /v1/device-status", authenticate(authenticator, handleClearStatus(statuses)))
	}
	return securityHeaders(mux)
}

func statusResponse(v store.StatusRequest) captureRequestResponse {
	return captureRequestResponse{RequestID: v.ID, Status: v.Status, CreatedAt: v.CreatedAt, ExpiresAt: v.ExpiresAt}
}

func handleStatusRequest(statuses StatusStore, events *eventHub, sender ...WakeupSender) authenticatedHandler {
	return func(w http.ResponseWriter, r *http.Request, deviceID string) {
		v, err := statuses.CreateStatusRequest(r.Context(), deviceID)
		if errors.Is(err, store.ErrPartnerNotFound) {
			writeError(w, 409, "partner_unavailable")
			return
		}
		if errors.Is(err, store.ErrRateLimited) {
			writeError(w, 429, "rate_limited")
			return
		}
		if err != nil {
			writeError(w, 500, "internal_error")
			return
		}
		events.publish(v.TargetDeviceID, deviceEvent{Type: "status.requested", RequestID: v.ID, ExpiresAt: v.ExpiresAt.Format(time.RFC3339)})
		if len(sender) > 0 && sender[0] != nil {
			if err := sender[0].SendWakeup(r.Context(), v.TargetDeviceID, v.PairID); err != nil {
				slog.Warn("FCM wakeup failed", "error", err)
			}
		}
		writeJSON(w, 201, statusResponse(v))
	}
}
func handlePendingStatusRequests(statuses StatusStore) authenticatedHandler {
	return func(w http.ResponseWriter, r *http.Request, deviceID string) {
		items, err := statuses.PendingStatusRequests(r.Context(), deviceID)
		if err != nil {
			writeError(w, 500, "internal_error")
			return
		}
		out := pendingCaptureResponse{Requests: make([]captureRequestResponse, 0, len(items))}
		for _, v := range items {
			out.Requests = append(out.Requests, statusResponse(v))
		}
		writeJSON(w, 200, out)
	}
}
func handleStatusResult(statuses StatusStore, events *eventHub) authenticatedHandler {
	return func(w http.ResponseWriter, r *http.Request, deviceID string) {
		var body statusResultRequest
		d := json.NewDecoder(http.MaxBytesReader(w, r.Body, 16*1024))
		d.DisallowUnknownFields()
		if d.Decode(&body) != nil {
			writeError(w, 400, "invalid_request")
			return
		}
		if body.Location.Status == "" {
			body.Location.Status = "DISABLED"
		}
		v, err := statuses.CompleteStatusRequest(r.Context(), deviceID, r.PathValue("requestID"), store.BatteryReport{Status: body.Battery.Status, Percent: body.Battery.Percent, ChargingState: body.Battery.ChargingState}, store.LocationReport{Status: body.Location.Status, Latitude: body.Location.Latitude, Longitude: body.Location.Longitude, AccuracyMeters: body.Location.AccuracyMeters, ObservedAt: body.Location.ObservedAt, Source: body.Location.Source})
		if errors.Is(err, store.ErrStatusRequestNotFound) {
			http.NotFound(w, r)
			return
		}
		if err != nil {
			writeError(w, 500, "internal_error")
			return
		}
		events.publish(v.RequesterDeviceID, deviceEvent{Type: "status.completed", RequestID: v.ID, Status: v.Status})
		w.WriteHeader(204)
	}
}
func handlePartnerStatus(statuses StatusStore) authenticatedHandler {
	return func(w http.ResponseWriter, r *http.Request, deviceID string) {
		v, err := statuses.PartnerStatusSnapshot(r.Context(), deviceID)
		if errors.Is(err, store.ErrStatusSnapshotNotFound) {
			http.NotFound(w, r)
			return
		}
		if err != nil {
			writeError(w, 500, "internal_error")
			return
		}
		writeJSON(w, 200, statusSnapshotResponse{DeviceID: v.DeviceID, Battery: batteryResultRequest{Status: v.Battery.Status, Percent: v.Battery.Percent, ChargingState: v.Battery.ChargingState}, Location: locationResultRequest{Status: v.Location.Status, Latitude: v.Location.Latitude, Longitude: v.Location.Longitude, AccuracyMeters: v.Location.AccuracyMeters, ObservedAt: v.Location.ObservedAt, Source: v.Location.Source}, ReportedAt: v.ReportedAt, ExpiresAt: v.ExpiresAt})
	}
}
func handleClearStatus(statuses StatusStore) authenticatedHandler {
	return func(w http.ResponseWriter, r *http.Request, deviceID string) {
		field := r.PathValue("field")
		if field == "" {
			field = "battery"
		}
		if field != "battery" && field != "location" {
			http.NotFound(w, r)
			return
		}
		if err := statuses.ClearStatusField(r.Context(), deviceID, field); err != nil {
			writeError(w, 500, "internal_error")
			return
		}
		w.WriteHeader(204)
	}
}

func handlePendingCaptureRequests(captures PendingCaptureStore) authenticatedHandler {
	return func(response http.ResponseWriter, request *http.Request, deviceID string) {
		requests, err := captures.PendingCaptureRequests(request.Context(), deviceID)
		if err != nil {
			writeError(response, http.StatusInternalServerError, "internal_error")
			return
		}
		result := pendingCaptureResponse{Requests: make([]captureRequestResponse, 0, len(requests))}
		for _, item := range requests {
			result.Requests = append(result.Requests, captureRequestResponse{RequestID: item.ID, Status: item.Status, CreatedAt: item.CreatedAt, ExpiresAt: item.ExpiresAt})
		}
		writeJSON(response, http.StatusOK, result)
	}
}
func handleCaptureRequestStatus(reader CaptureReader) authenticatedHandler {
	return func(w http.ResponseWriter, r *http.Request, deviceID string) {
		v, err := reader.CaptureRequestForRequester(r.Context(), deviceID, r.PathValue("requestID"))
		if errors.Is(err, store.ErrCaptureRequestNotFound) {
			http.NotFound(w, r)
			return
		}
		if err != nil {
			writeError(w, 500, "internal_error")
			return
		}
		writeJSON(w, 200, struct {
			RequestID string    `json:"requestId"`
			Status    string    `json:"status"`
			ImageID   string    `json:"imageId"`
			Failure   string    `json:"failure"`
			ExpiresAt time.Time `json:"expiresAt"`
		}{v.ID, v.Status, v.ImageID, v.Failure, v.ExpiresAt})
	}
}

type authenticatedHandler func(http.ResponseWriter, *http.Request, string)

func authenticate(authenticator Authenticator, next authenticatedHandler) http.HandlerFunc {
	return func(response http.ResponseWriter, request *http.Request) {
		authorization := request.Header.Get("Authorization")
		if !strings.HasPrefix(authorization, "Bearer ") || strings.Contains(strings.TrimPrefix(authorization, "Bearer "), " ") {
			writeError(response, http.StatusUnauthorized, "unauthorized")
			return
		}
		deviceID, err := authenticator.AuthenticateDevice(request.Context(), strings.TrimPrefix(authorization, "Bearer "))
		if errors.Is(err, store.ErrUnauthorized) {
			writeError(response, http.StatusUnauthorized, "unauthorized")
			return
		}
		if err != nil {
			writeError(response, http.StatusInternalServerError, "internal_error")
			return
		}
		next(response, request, deviceID)
	}
}

func handleEvents(authenticator Authenticator, events *eventHub) http.HandlerFunc {
	return func(response http.ResponseWriter, request *http.Request) {
		authorization := request.Header.Get("Authorization")
		if !strings.HasPrefix(authorization, "Bearer ") {
			writeError(response, http.StatusUnauthorized, "unauthorized")
			return
		}
		deviceID, err := authenticator.AuthenticateDevice(request.Context(), strings.TrimPrefix(authorization, "Bearer "))
		if err != nil {
			writeError(response, http.StatusUnauthorized, "unauthorized")
			return
		}
		connection, err := websocket.Accept(response, request, &websocket.AcceptOptions{
			CompressionMode: websocket.CompressionDisabled,
		})
		if err != nil {
			slog.Warn("websocket upgrade failed")
			return
		}
		slog.Info("websocket connected", "device_id", deviceID)
		defer func() { _ = connection.Close(websocket.StatusNormalClosure, "") }()
		connection.SetReadLimit(16 * 1024)
		ctx := connection.CloseRead(request.Context())
		channel, unsubscribe := events.subscribe(deviceID)
		defer unsubscribe()
		defer slog.Info("websocket disconnected", "device_id", deviceID)
		for {
			select {
			case <-ctx.Done():
				return
			case event := <-channel:
				writeCtx, cancel := context.WithTimeout(ctx, 5*time.Second)
				err := wsjson.Write(writeCtx, connection, event)
				cancel()
				if err != nil {
					return
				}
			}
		}
	}
}

func handleCaptureRequest(captures CaptureStore, events *eventHub, sender ...WakeupSender) authenticatedHandler {
	return func(response http.ResponseWriter, request *http.Request, deviceID string) {
		capture, err := captures.CreateCaptureRequest(request.Context(), deviceID)
		if errors.Is(err, store.ErrPartnerNotFound) {
			writeError(response, http.StatusConflict, "partner_unavailable")
			return
		}
		if errors.Is(err, store.ErrRateLimited) {
			writeError(response, http.StatusTooManyRequests, "rate_limited")
			return
		}
		if err != nil {
			writeError(response, http.StatusInternalServerError, "internal_error")
			return
		}
		events.publish(capture.TargetDeviceID, deviceEvent{
			Type: "capture.requested", RequestID: capture.ID, ExpiresAt: capture.ExpiresAt.Format(time.RFC3339),
		})
		if len(sender) > 0 && sender[0] != nil {
			if err := sender[0].SendWakeup(request.Context(), capture.TargetDeviceID, capture.PairID); err != nil {
				slog.Warn("FCM wakeup failed", "error", err)
			}
		}
		writeJSON(response, http.StatusCreated, captureRequestResponse{
			RequestID: capture.ID, Status: capture.Status, CreatedAt: capture.CreatedAt, ExpiresAt: capture.ExpiresAt,
		})
	}
}

func handleFCMToken(tokens FCMTokenStore) authenticatedHandler {
	return func(response http.ResponseWriter, request *http.Request, deviceID string) {
		request.Body = http.MaxBytesReader(response, request.Body, 8*1024)
		var body fcmTokenRequest
		decoder := json.NewDecoder(request.Body)
		decoder.DisallowUnknownFields()
		if err := decoder.Decode(&body); err != nil || strings.TrimSpace(body.Token) == "" || len(body.Token) > 4096 {
			writeError(response, http.StatusBadRequest, "invalid_request")
			return
		}
		if err := tokens.SetFCMToken(request.Context(), deviceID, strings.TrimSpace(body.Token)); err != nil {
			writeError(response, http.StatusInternalServerError, "internal_error")
			return
		}
		response.WriteHeader(http.StatusNoContent)
	}
}

func handleCaptureResult(captures CaptureStore, events *eventHub) authenticatedHandler {
	return func(response http.ResponseWriter, request *http.Request, deviceID string) {
		request.Body = http.MaxBytesReader(response, request.Body, 16*1024)
		decoder := json.NewDecoder(request.Body)
		decoder.DisallowUnknownFields()
		var body captureResultRequest
		if err := decoder.Decode(&body); err != nil {
			writeError(response, http.StatusBadRequest, "invalid_request")
			return
		}
		if err := decoder.Decode(&struct{}{}); !errors.Is(err, io.EOF) || !validCaptureResult(body) || request.PathValue("requestID") == "" {
			writeError(response, http.StatusBadRequest, "invalid_request")
			return
		}
		capture, err := captures.CompleteCaptureRequest(
			request.Context(), deviceID, request.PathValue("requestID"), body.Status, body.ImageID, body.Failure,
		)
		if errors.Is(err, store.ErrCaptureRequestNotFound) {
			http.NotFound(response, request)
			return
		}
		if err != nil {
			writeError(response, http.StatusInternalServerError, "internal_error")
			return
		}
		events.publish(capture.RequesterDeviceID, deviceEvent{
			Type: "capture.completed", RequestID: capture.ID, Status: capture.Status,
			ImageID: capture.ImageID, Failure: capture.Failure,
		})
		response.WriteHeader(http.StatusNoContent)
	}
}

func validCaptureResult(body captureResultRequest) bool {
	if body.Status == "READY" {
		return body.ImageID != "" && body.Failure == "" && !strings.ContainsAny(body.ImageID, "/\\")
	}
	if body.Status != "FAILED" || body.ImageID != "" {
		return false
	}
	for _, allowed := range []string{"DISABLED", "SERVICE_UNAVAILABLE", "LOCKED", "CAPTURE_PROTECTED", "INTERNAL_ERROR"} {
		if body.Failure == allowed {
			return true
		}
	}
	return false
}

func handleImageUpload(images ImageStore) authenticatedHandler {
	return func(response http.ResponseWriter, request *http.Request, deviceID string) {
		if request.Header.Get("Content-Type") != "image/jpeg" {
			writeError(response, http.StatusUnsupportedMediaType, "jpeg_required")
			return
		}
		data, err := io.ReadAll(io.LimitReader(request.Body, maxImageBytes+1))
		if err != nil || len(data) == 0 || len(data) > maxImageBytes {
			writeError(response, http.StatusRequestEntityTooLarge, "image_too_large")
			return
		}
		config, format, err := image.DecodeConfig(bytes.NewReader(data))
		if err != nil || format != "jpeg" || config.Width <= 0 || config.Height <= 0 || config.Width > maxImagePixels/config.Height {
			writeError(response, http.StatusBadRequest, "invalid_image")
			return
		}
		stored, err := images.SaveImage(request.Context(), deviceID, data, config.Width, config.Height)
		if err != nil {
			writeError(response, http.StatusInternalServerError, "internal_error")
			return
		}
		writeJSON(response, http.StatusCreated, imageResponse{ImageID: stored.ID, CreatedAt: stored.CreatedAt, ExpiresAt: stored.ExpiresAt})
	}
}

func handleImageDownload(images ImageStore) authenticatedHandler {
	return func(response http.ResponseWriter, request *http.Request, deviceID string) {
		imageID := request.PathValue("imageID")
		if imageID == "" || strings.ContainsAny(imageID, "/\\") {
			http.NotFound(response, request)
			return
		}
		stored, err := images.TakeImage(request.Context(), deviceID, imageID)
		if errors.Is(err, store.ErrImageNotFound) {
			http.NotFound(response, request)
			return
		}
		if err != nil {
			writeError(response, http.StatusInternalServerError, "internal_error")
			return
		}
		response.Header().Set("Content-Type", "image/jpeg")
		response.Header().Set("Content-Length", fmt.Sprint(len(stored.Data)))
		response.WriteHeader(http.StatusOK)
		_, _ = response.Write(stored.Data)
	}
}

func handleHealth(response http.ResponseWriter, _ *http.Request) {
	response.Header().Set("Content-Type", "application/json")
	response.WriteHeader(http.StatusOK)
	_ = json.NewEncoder(response).Encode(healthResponse{Status: "ok"})
}

func handleEnrollment(enroller Enroller) http.HandlerFunc {
	return func(response http.ResponseWriter, request *http.Request) {
		if enroller == nil {
			writeError(response, http.StatusInternalServerError, "internal_error")
			return
		}
		request.Body = http.MaxBytesReader(response, request.Body, 16*1024)
		decoder := json.NewDecoder(request.Body)
		decoder.DisallowUnknownFields()

		var body enrollmentRequest
		if err := decoder.Decode(&body); err != nil {
			writeError(response, http.StatusBadRequest, "invalid_request")
			return
		}
		if err := decoder.Decode(&struct{}{}); !errors.Is(err, io.EOF) {
			writeError(response, http.StatusBadRequest, "invalid_request")
			return
		}
		body.DeviceName = strings.TrimSpace(body.DeviceName)
		if !validEnrollment(body) {
			writeError(response, http.StatusBadRequest, "invalid_request")
			return
		}

		enrollment, err := enroller.EnrollDevice(
			request.Context(),
			body.InvitationToken,
			body.DeviceName,
			body.PublicKey,
		)
		if errors.Is(err, store.ErrInvitationNotFound) {
			http.NotFound(response, request)
			return
		}
		if err != nil {
			writeError(response, http.StatusInternalServerError, "internal_error")
			return
		}

		writeJSON(response, http.StatusCreated, enrollmentResponse{
			DeviceID:   enrollment.DeviceID,
			PairID:     enrollment.PairID,
			Slot:       enrollment.Slot,
			Credential: enrollment.Credential,
		})
	}
}

func validEnrollment(request enrollmentRequest) bool {
	if len(request.InvitationToken) < 43 || len(request.InvitationToken) > 128 {
		return false
	}
	if !utf8.ValidString(request.DeviceName) || utf8.RuneCountInString(request.DeviceName) < 1 || utf8.RuneCountInString(request.DeviceName) > 80 {
		return false
	}
	publicKeyBytes, err := base64.RawURLEncoding.DecodeString(request.PublicKey)
	if err != nil {
		return false
	}
	parsed, err := x509.ParsePKIXPublicKey(publicKeyBytes)
	publicKey, ok := parsed.(*ecdsa.PublicKey)
	return err == nil && ok && publicKey.Curve == elliptic.P256()
}

func writeError(response http.ResponseWriter, status int, code string) {
	writeJSON(response, status, errorResponse{Code: code})
}

func writeJSON(response http.ResponseWriter, status int, value any) {
	response.Header().Set("Content-Type", "application/json")
	response.WriteHeader(status)
	_ = json.NewEncoder(response).Encode(value)
}

func securityHeaders(next http.Handler) http.Handler {
	return http.HandlerFunc(func(response http.ResponseWriter, request *http.Request) {
		response.Header().Set("Cache-Control", "no-store")
		response.Header().Set("X-Content-Type-Options", "nosniff")
		next.ServeHTTP(response, request)
	})
}